changed byte[] to int[] and float[], in java classes. removed jvm-opts, as these don't work with cake.

+++ src/jvm/forma/FloatsSerialization.java
@@ -108,7 +108,7 @@ import org.apache.hadoop.io.serializer.Serializer;
         }
 
     @Override
-        public Comparator<byte[]> getComparator( Class<byte[]> type )
+        public Comparator<float[]> getComparator( Class<float[]> type )
         {
             return new BytesComparator();
         }
